refactor: Move LinkingDirection to o.o.street.model.edge
The enum is used in StreetEdge, so it need to be in the model
to avoid circular dependencies
