+++ application/src/test/java/org/opentripplanner/graph_builder/module/linking/TestGraph.java
@@ -2,8 +2,8 @@ package org.opentripplanner.graph_builder.module.linking;
 
 import java.util.List;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.routing.linking.LinkingDirection;
 import org.opentripplanner.routing.linking.VertexLinker;
+import org.opentripplanner.street.model.edge.LinkingDirection;
 import org.opentripplanner.street.model.edge.StreetTransitStopLink;
 import org.opentripplanner.street.model.vertex.TransitStopVertex;
 import org.opentripplanner.street.search.TraverseMode;
