+++ application/src/test/java/org/opentripplanner/routing/linking/LinkingDirectionTest.java
@@ -4,6 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
+import org.opentripplanner.street.model.edge.LinkingDirection;
 
 class LinkingDirectionTest {
